+++ src/components/index.js
@@ -10,6 +10,7 @@ import Button from './Widgets/Button'
 import Checkbox from './Widgets/Checkbox'
 import RadioButton from './Widgets/Form/RadioButton'
 import Input from './Widgets/Form/Input'
+import AddressInput from './Widgets/Form/AddressInput'
 import TextArea from './Widgets/Form/TextArea'
 import Select from './Widgets/Form/Select'
 import FileChooser from './Widgets/Form/FileChooser'
@@ -17,18 +18,22 @@ import ValidatedField from './Widgets/Form/ValidatedField'
 import ConverterForm from './Tools/EthConverterForm'
 import NetworkChooser from './Network/NetworkChooser'
 import RpcTester from './Network/RPC/RpcTester'
-import TxHistory from './Tx/History'
 import FormSendTx from './Tx/SendTx'
+import TxHistory from './Tx/TxHistory'
 
 import AccountList from './Wallet/AccountList'
 import AccountItem from './Wallet/AccountItem'
 import TokenListForItem from './Wallet/TokenListForItem'
 import NavbarItem from './Wallet/NavbarItem'
 import NetworkStatus from './Wallet/NetworkStatus'
+import Notification from './Wallet/Notification'
+
+import * as utils from '../lib/util'
 
 export {
   AccountList,
   AccountItem,
+  AddressInput,
   Button,
   Checkbox,
   RadioButton,
@@ -48,10 +53,12 @@ export {
   NetworkStatus,
   NodeInfo,
   NodeInfoBox,
+  Notification,
   Pulse,
   RpcTester,
   Spinner,
   TokenListForItem,
   TxHistory,
-  ValidatedField
+  ValidatedField,
+  utils
 }